Installed db from repository, updated WP settings.
Note that there's still a number of templates to be created, but we're
both on the same page now.
I deactivated the eshop plugin as I thought it was causing problems,
turns out it was because you'd set the index page to be a specific page
(for which a template does not yet exist. The plugins can be
reactivated whenever.